+++ app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".authentication.EmailPasswordActivity">
+    tools:context=".authentication.LoginActivity">
 
     <!-- Login progress -->
     <ProgressBar
