+++ app/src/main/res/layout/activity_registration.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".authentication.EmailPasswordActivity">
+    tools:context=".authentication.RegistrationActivity">
 
     <!-- Login progress -->
     <ProgressBar
@@ -98,7 +98,7 @@
             </android.support.design.widget.TextInputLayout>
 
             <Button
-                android:id="@+id/btnRegister"
+                android:id="@+id/email_create_account_button"
                 style="?android:textAppearanceSmall"
                 android:layout_width="100dp"
                 android:layout_height="wrap_content"
